pyproject.toml
[project]
name = "drfproject"
version = "0.1.0"
description = ""
authors = [
    {name = "Your Name",email = "you@example.com"}
]
readme = "README.md"
requires-python = ">=3.12,<4.0"
dependencies = [
    "django (>=5.2.7,<6.0.0)",
    "psycopg2 (>=2.9.11,<3.0.0)",
    "pillow (>=12.0.0,<13.0.0)",
    "djangorestframework (>=3.16.1,<4.0.0)",
    "dotenv (>=0.9.9,<0.10.0)",
    "django-filter (>=25.2,<26.0)",
    "djangorestframework-simplejwt (>=5.5.1,<6.0.0)",
    "drf-yasg (>=1.21.11,<2.0.0)",
    "django-cors-headers (>=4.9.0,<5.0.0)",
    "requests (>=2.32.5,<3.0.0)",
    "stripe (>=13.2.0,<14.0.0)",
    "celery (>=5.5.3,<6.0.0)",
    "eventlet (>=0.40.3,<0.41.0)",
    "redis (>=7.0.1,<8.0.0)",
    "django-celery-beat (>=2.8.1,<3.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
